Fixed template plugin tests
